Added logback
This commit adds logback which allows us to log things to both the console and a file at the same time. This also allows us to have a little more control over what we see from the Discord library's logging.

This is the last thing on the release milestone! As long as this version seems stable, we can release Plexi V1.0!